+++ wow_login_messages/src/util/async_std_impl.rs
@@ -81,3 +81,11 @@ pub async fn astd_read_u64_le<R: ReadExt + Unpin>(mut r: R) -> Result<u64, std::
     r.read_exact(&mut v).await?;
     Ok(u64::from_le_bytes(v))
 }
+
+// i32
+pub async fn astd_read_i32_le<R: ReadExt + Unpin>(mut r: R) -> Result<i32, std::io::Error> {
+    let mut v = [0_u8; 4];
+    r.read_exact(&mut v).await?;
+    Ok(i32::from_le_bytes(v))
+}
+